requirements.txt
mysql-connector-python
pyqt6